Revamp API
API now resembles that of SciML.jl and Julia's sorting utilities.

There is now a single·user-facing command, `sinfit`. This command
performs a regression on a `Problem` using the specified `Algorithm`.
The `Problem` type is used to specify the type of regression (plain
sinusoid, mixed linear-sinusoid, etc), along with all the data that
is available (sample times, samples, initial estimates, bounds...).
The `Algorithm` type specifies which algorithm to use to perform
the regression (ieee1057, Jacquelin, non-linear LS...).